requirements.txt
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
deep-translator==1.11.4
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
pycparser==3.0
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
